Adds additional hint.
There's an idea I've been made aware of called undesirable difficulty in learning theory. The idea is that sometimes obstacles are put in place that take the focus off learning a specific thing by introducing a challenge on a thing that might be related but not critical to the task at hand.

In this exercise, I notice pretty much everyone reaches for the splice method and I suspect it is because of the language in the prompt. Specifically: "removes the other arguments from that array"

I've added a hint that I think would not take away from the challenge but give learners an alternative way to think about the exercise.